+++ baybe/searchspace/continuous.py
@@ -13,6 +13,8 @@ from attr import define, field
 
 from baybe.constraints import (
     ContinuousCardinalityConstraint,
+    ContinuousInterPointLinearEqualityConstraint,
+    ContinuousInterPointLinearInequalityConstraint,
     ContinuousLinearEqualityConstraint,
     ContinuousLinearInequalityConstraint,
 )
@@ -68,6 +70,16 @@ class SubspaceContinuous(SerialMixin):
     )
     """Nonlinear constraints."""
 
+    constraints_interpoint_lin_eq: tuple[
+        ContinuousInterPointLinearEqualityConstraint
+    ] = field(converter=to_tuple, factory=tuple)
+    "Interpoint linear equality constraints."
+
+    constraints_interpoint_lin_ineq: tuple[
+        ContinuousInterPointLinearInequalityConstraint
+    ] = field(converter=to_tuple, factory=tuple)
+    "Interpoint linear inequality constraints."
+
     def __str__(self) -> str:
         if self.is_empty:
             return ""
@@ -81,15 +93,34 @@ class SubspaceContinuous(SerialMixin):
         nonlin_constraints_list = [
             constr.summary() for constr in self.constraints_nonlin
         ]
+        interpoint_eq_constraints_list = [
+            constr.summary() for constr in self.constraints_interpoint_lin_eq
+        ]
+        interpoint_ineq_constraints_list = [
+            constr.summary() for constr in self.constraints_interpoint_lin_ineq
+        ]
+        nonlin_constraints_list = [
+            constr.summary() for constr in self.constraints_nonlin
+        ]
         param_df = pd.DataFrame(param_list)
         lin_eq_df = pd.DataFrame(eq_constraints_list)
         lin_ineq_df = pd.DataFrame(ineq_constraints_list)
+        interpoint_lin_eq_df = pd.DataFrame(interpoint_eq_constraints_list)
+        interpoint_lin_ineq_df = pd.DataFrame(interpoint_ineq_constraints_list)
         nonlinear_df = pd.DataFrame(nonlin_constraints_list)
 
         fields = [
             to_string("Continuous Parameters", pretty_print_df(param_df)),
             to_string("Linear Equality Constraints", pretty_print_df(lin_eq_df)),
             to_string("Linear Inequality Constraints", pretty_print_df(lin_ineq_df)),
+            to_string(
+                "Linear Interpoint Equality Constraints",
+                pretty_print_df(interpoint_lin_eq_df),
+            ),
+            to_string(
+                "Linear Interpoint Inequality Constraints",
+                pretty_print_df(interpoint_lin_ineq_df),
+            ),
             to_string("Non-linear Constraints", pretty_print_df(nonlinear_df)),
         ]
 
@@ -154,6 +185,16 @@ class SubspaceContinuous(SerialMixin):
                 c
                 for c in constraints
                 if isinstance(c, ContinuousLinearInequalityConstraint)
+            ],
+            constraints_interpoint_lin_eq=[  # type:ignore[misc]
+                c
+                for c in constraints
+                if isinstance(c, ContinuousInterPointLinearEqualityConstraint)
+            ],
+            constraints_interpoint_lin_ineq=[  # type:ignore[misc]
+                c
+                for c in constraints
+                if isinstance(c, ContinuousInterPointLinearInequalityConstraint)
             ],
             constraints_nonlin=[
                 c for c in constraints if isinstance(c, ContinuousNonlinearConstraint)
@@ -253,6 +294,26 @@ class SubspaceContinuous(SerialMixin):
         return pd.DataFrame(
             {p.name: p.bounds.to_tuple() for p in self.parameters},
             index=["min", "max"],
+        )
+
+    @property
+    def is_constrained(self) -> bool:
+        """Return whether the subspace is constrained in any way."""
+        return any(
+            (
+                self.constraints_interpoint_lin_eq,
+                self.constraints_interpoint_lin_ineq,
+                self.constraints_lin_eq,
+                self.constraints_lin_ineq,
+                self.constraints_nonlin,
+            )
+        )
+
+    @property
+    def has_interpoint_constraints(self) -> bool:
+        """Return whether or not the space has any interpoint constraints."""
+        return any(
+            (self.constraints_interpoint_lin_eq, self.constraints_interpoint_lin_ineq)
         )
 
     def _drop_parameters(self, parameter_names: Collection[str]) -> SubspaceContinuous:
@@ -356,14 +417,16 @@ class SubspaceContinuous(SerialMixin):
 
         if not self.parameters:
             return pd.DataFrame(index=pd.RangeIndex(0, batch_size))
-
-        if (
-            len(self.constraints_lin_eq) == 0
-            and len(self.constraints_lin_ineq) == 0
-            and len(self.constraints_cardinality) == 0
-        ):
+        # If the space is completely unconstrained, we can sample from bounds.
+        if not self.is_constrained:
             return self._sample_from_bounds(batch_size, self.comp_rep_bounds.values)
 
+        if self.has_interpoint_constraints:
+            return self._sample_from_polytope_with_interpoint_constraints(
+                batch_size, self.comp_rep_bounds.values
+            )
+
+        # If there are no cardinality constraints, we sample directly from the polytope
         if len(self.constraints_cardinality) == 0:
             return self._sample_from_polytope(batch_size, self.comp_rep_bounds.values)
 
@@ -375,6 +438,146 @@ class SubspaceContinuous(SerialMixin):
             low=bounds[0, :], high=bounds[1, :], size=(batch_size, len(self.parameters))
         )
 
+        return pd.DataFrame(points, columns=self.parameter_names)
+
+    def _sample_from_polytope_with_interpoint_constraints(
+        self, batch_size: int, bounds: np.ndarray
+    ) -> pd.DataFrame:
+        """Draw uniform random samples from a polytope with interpoint constraints."""
+        # If the space has interpoint constraints, we need to sample from a larger
+        # searchspace that models the batch size via additional dimension. This is
+        # necessary since `get_polytope_samples` cannot handle inter-point constraints,
+        # see https://github.com/pytorch/botorch/issues/2468
+
+        import torch
+        from botorch.utils.sampling import get_polytope_samples
+
+        from baybe.utils.numerical import DTypeFloatNumpy
+        from baybe.utils.torch import DTypeFloatTorch
+
+        # The number of parameters is needed at some places for adjusting indices
+        num_of_params = len(self.parameters)
+
+        eq_constraints, ineq_constraints = [], []
+        # We start with creating the interpoint constraints
+        for c in [
+            *self.constraints_interpoint_lin_eq,
+            *self.constraints_interpoint_lin_ineq,
+        ]:
+            # Get the index of the parameter that is used in this constraint
+            param_offset = self.parameter_names.index(c.parameters[0])
+            if batch_size < max(c._batchnumbers) + 1:
+                raise ValueError(
+                    f"Number of requested batches is smaller than number of "
+                    f"batches defined in the constraint {c}"
+                )
+            param_indices = [
+                batch_number * num_of_params + param_offset
+                for batch_number in c._batchnumbers
+            ]
+            coefficients = c._raw_coefficients
+            botorch_tuple = (
+                torch.tensor(param_indices),
+                torch.tensor(coefficients, dtype=DTypeFloatTorch),
+                np.asarray(c.rhs, dtype=DTypeFloatNumpy).item(),
+            )
+            if isinstance(c, ContinuousInterPointLinearEqualityConstraint):
+                eq_constraints.append(botorch_tuple)
+            else:
+                ineq_constraints.append(botorch_tuple)
+
+        # We now continue with the non-interpoint constraints which also need to be
+        # adjusted.
+        for c in [*self.constraints_lin_eq, *self.constraints_lin_ineq]:
+            param_indices, coefficients, rhs = c.to_botorch(self.parameters)
+            for b in range(batch_size):
+                botorch_tuple = (param_indices + b * num_of_params, coefficients, rhs)
+                if isinstance(c, ContinuousLinearEqualityConstraint):
+                    eq_constraints.append(botorch_tuple)
+                else:
+                    ineq_constraints.append(botorch_tuple)
+
+        bounds_joint = torch.cat(
+            [torch.from_numpy(bounds) for _ in range(batch_size)], dim=-1
+        )
+        points = get_polytope_samples(
+            n=1,
+            bounds=bounds_joint,
+            equality_constraints=eq_constraints,
+            inequality_constraints=ineq_constraints,
+        )
+        points = points.reshape(batch_size, points.shape[-1] // batch_size)
+        return pd.DataFrame(points, columns=self.parameter_names)
+
+    def _sample_from_polytope_with_interpoint_constraints(
+        self, batch_size: int, bounds: np.ndarray
+    ) -> pd.DataFrame:
+        """Draw uniform random samples from a polytope with interpoint constraints."""
+        # If the space has interpoint constraints, we need to sample from a larger
+        # searchspace that models the batch size via additional dimension. This is
+        # necessary since `get_polytope_samples` cannot handle inter-point constraints,
+        # see https://github.com/pytorch/botorch/issues/2468
+
+        import torch
+        from botorch.utils.sampling import get_polytope_samples
+
+        from baybe.utils.numerical import DTypeFloatNumpy
+        from baybe.utils.torch import DTypeFloatTorch
+
+        # The number of parameters is needed at some places for adjusting indices
+        num_of_params = len(self.parameters)
+
+        eq_constraints, ineq_constraints = [], []
+        # We start with creating the interpoint constraints
+        for c in [
+            *self.constraints_interpoint_lin_eq,
+            *self.constraints_interpoint_lin_ineq,
+        ]:
+            # Get the offset of the parameters used in this constraint
+            param_offset = {
+                name: self.parameter_names.index(name) for name in c.parameters
+            }
+            if batch_size < max(c._batchnumbers) + 1:
+                raise ValueError(
+                    f"Number of requested batches is smaller than number of "
+                    f"batches defined in the constraint {c}"
+                )
+            param_indices = [
+                batch * num_of_params + param_offset[name]
+                for name, batch, _ in c.coefficients
+            ]
+            coefficients = c._raw_coefficients
+            botorch_tuple = (
+                torch.tensor(param_indices),
+                torch.tensor(coefficients, dtype=DTypeFloatTorch),
+                np.asarray(c.rhs, dtype=DTypeFloatNumpy).item(),
+            )
+            if isinstance(c, ContinuousInterPointLinearEqualityConstraint):
+                eq_constraints.append(botorch_tuple)
+            else:
+                ineq_constraints.append(botorch_tuple)
+
+        # We now continue with the non-interpoint constraints which also need to be
+        # adjusted.
+        for c in [*self.constraints_lin_eq, *self.constraints_lin_ineq]:
+            param_indices, coefficients, rhs = c.to_botorch(self.parameters)
+            for b in range(batch_size):
+                botorch_tuple = (param_indices + b * num_of_params, coefficients, rhs)
+                if isinstance(c, ContinuousLinearEqualityConstraint):
+                    eq_constraints.append(botorch_tuple)
+                else:
+                    ineq_constraints.append(botorch_tuple)
+
+        bounds_joint = torch.cat(
+            [torch.from_numpy(bounds) for _ in range(batch_size)], dim=-1
+        )
+        points = get_polytope_samples(
+            n=1,
+            bounds=bounds_joint,
+            equality_constraints=eq_constraints,
+            inequality_constraints=ineq_constraints,
+        )
+        points = points.reshape(batch_size, points.shape[-1] // batch_size)
         return pd.DataFrame(points, columns=self.parameter_names)
 
     def _sample_from_polytope(