+++ examples/Constraints_Continuous/linear_constraints.py
@@ -19,6 +19,8 @@ from botorch.test_functions import Rastrigin
 
 from baybe import Campaign
 from baybe.constraints import (
+    ContinuousInterPointLinearEqualityConstraint,
+    ContinuousInterPointLinearInequalityConstraint,
     ContinuousLinearEqualityConstraint,
     ContinuousLinearInequalityConstraint,
 )
@@ -92,7 +94,7 @@ campaign = Campaign(
 
 SMOKE_TEST = "SMOKE_TEST" in os.environ
 
-BATCH_SIZE = 2 if SMOKE_TEST else 3
+BATCH_SIZE = 4 if SMOKE_TEST else 5
 N_ITERATIONS = 2 if SMOKE_TEST else 3
 
 for k in range(N_ITERATIONS):
@@ -143,3 +145,86 @@ print(
     "2.0*x_2 + 3.0*x_4 <= 1.0 satisfied in all recommendations? ",
     (2.0 * measurements["x_2"] + 3.0 * measurements["x_4"]).le(1.0 + TOLERANCE).all(),
 )
+
+
+### Using inter-point constraints
+
+# It is also possible to require inter-point constraints which constraint the value of
+# a single parameter across a full batch.
+# This example models the following inter-point constraints and combines them also
+# with regular constraints.
+# 1. The sum of `x_1` across the first three batches needs to be at least 2.5.
+# 2. The sum of `x_2` across the first four batches needs to be exactly 3.
+# 3. The sum of `x_1` and `x_3` across the first three batches needs to be at least 5.
+
+inter_constraints = [
+    ContinuousInterPointLinearInequalityConstraint(
+        parameters=["x_1"],
+        coefficients=[("x_1", 0, 1), ("x_1", 1, 1), ("x_1", 2, 1)],
+        rhs=2.5,
+    ),
+    ContinuousInterPointLinearEqualityConstraint(
+        parameters="x_2",  # For single parameters, parameter names are converted.
+        coefficients=[("x_2", 0, 1), ("x_2", 1, 1), ("x_2", 2, 1), ("x_2", 3, 1)],
+        rhs=3,
+    ),
+    ContinuousInterPointLinearEqualityConstraint(
+        parameters=["x_1", "x_3"],
+        coefficients=[
+            ("x_1", 0, 1),
+            ("x_1", 1, 1),
+            ("x_1", 2, 1),
+            ("x_3", 0, 1),
+            ("x_3", 1, 1),
+            ("x_3", 2, 1),
+        ],
+        rhs=5,
+    ),
+    ContinuousLinearEqualityConstraint(
+        parameters=["x_1", "x_2"], coefficients=[1, 1], rhs=2
+    ),
+    ContinuousLinearInequalityConstraint(
+        parameters=["x_3", "x_4"], coefficients=[-1, 1], rhs=-1
+    ),
+]
+
+### Re-construct search space and campaign and run some iterations
+
+inter_searchspace = SearchSpace.from_product(
+    parameters=parameters, constraints=inter_constraints
+)
+
+inter_campaign = Campaign(
+    searchspace=inter_searchspace,
+    objective=objective,
+)
+
+for k in range(N_ITERATIONS):
+    rec = inter_campaign.recommend(batch_size=BATCH_SIZE)
+
+    # target value are looked up via the botorch wrapper
+    target_values = []
+    for index, row in rec.iterrows():
+        target_values.append(WRAPPED_FUNCTION(*row.to_list()))
+
+    rec["Target"] = target_values
+    inter_campaign.add_measurements(rec)
+    # Check inter-point constraints
+    assert rec.at[0, "x_1"] + rec.at[1, "x_1"] + rec.at[2, "x_1"] >= 2.5 - TOLERANCE
+    assert np.isclose(
+        rec.at[0, "x_2"] + rec.at[1, "x_2"] + rec.at[2, "x_2"] + rec.at[3, "x_2"],
+        3,
+    )
+    assert (
+        rec.at[0, "x_1"]
+        + rec.at[1, "x_1"]
+        + rec.at[2, "x_1"]
+        + rec.at[0, "x_3"]
+        + rec.at[1, "x_3"]
+        + rec.at[2, "x_3"]
+        >= 4 - TOLERANCE
+    )
+    # Check normal constraints for each batch individually
+    for b in range(BATCH_SIZE):
+        assert np.isclose(rec.at[b, "x_1"] + rec.at[b, "x_2"], 2)
+        assert -rec.at[b, "x_3"] + rec.at[b, "x_4"] >= -1 - TOLERANCE